Remove directory generator
